# Speed up the module cache 6x by removing LZW compression